requirements.txt

firecrawl-py==0.0.20
langchain==0.3.7
langchain-openai==0.2.8
langgraph==0.2.34
pydantic==2.9.2
langchain-mcp-adapters==0.1.2
python-dotenv==1.0.1
fastapi==0.115.4
uvicorn[standard]==0.32.0
jinja2==3.1.4